Create FM_Master.psc
DESCRIPTION: A mod that allows the player to edit the relationship between the player and wolves (members of the wolf faction). By default, if the player is a werewolf, wolves only have a disposition of 2 with the player, meaning that they are merely confidant in the player (will not attack player unless provoked). This mod allows the player to change this value to whatever they see fit, and even set separate values depending on whether or not the player is a werewolf. Furthermore,
; the user can even change the relationship of wolves with other factions, such as
; vampires. I for one find it somewhat strange that wolves are more friendly towards vampires than they are towards werewolves, and wrote this script to correct this.

PLANS FOR FUTURE VERSIONS: In the future, I plan to allow players to approach any wolf in the wild and make them a follower of the player. I also plan to create an option that allows the player to summon a pack of wolves upon transforming into a werewolf.